deleting responsable column only if it exists - added check

+++ database/migrations/2022_04_29_141158_remove_responsable_column_from_services.php
@@ -13,9 +13,13 @@ class RemoveResponsableColumnFromServices extends Migration
      */
     public function up()
     {
-        Schema::table('services', function (Blueprint $table) {
-            $table->dropColumn('responsable');
-        });
+        if (Schema::hasColumn('services', 'responsable'))
+        {
+            Schema::table('services', function (Blueprint $table)
+            {
+                $table->dropColumn('responsable');
+            });
+        }
     }
 
     /**
